--- src/features/forceUpdate/__mocks__/useMustUpdateApp.ts
@@ -1,1 +1,0 @@
-export const useMustUpdateApp = jest.fn().mockReturnValue(false)
