+++ src/features/forceUpdate/pages/ForceUpdate.native.test.tsx
@@ -2,7 +2,9 @@ import React from 'react'
 
 import { render } from 'tests/utils'
 
-import { ForceUpdate } from '../ForceUpdate'
+import { ForceUpdate } from './ForceUpdate'
+
+export const useMustUpdateApp = jest.fn().mockReturnValue(false)
 
 describe('<ForceUpdate/>', () => {
   it('should match snapshot', async () => {
